Add handle_cast tracing to get times for how long event processing takes.
This is done by using the 'dbg' lib to remotely trace the calls to the
handle_cast/2 function in the stats db process inside Rabbit MQ.
For each of the incoming events, the time taken to handle them is tallied
up. Allowing a breakdown of the time taken by the stats db server per event
type.